Fix bug when parsing nested subcommands without positional arguments for each of them.
The bug caused an assertion to fail exiting the parse loop. When the loop finishes with more than one
element in the subcommand stack, only the topmost is popped yielding the failure.
Fix by making sure that upon exit from the parse loop all remaining subcommands in the stack are
finalized.